Update the sidebar on all the pages
I changed the sidebar on the website to say
"Mythic Roles" instead of saying "What does each
mythic do" so that the scope of the guide itself
would be smaller and more easily digestible.

I think it would be easier to just add information
specific to each mythic on a page dedicated to
explaining every mythic.